requirements.txt
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1.post1
pyyaml==6.0.1
requests==2.31.0
pytest==8.1.1